moves setupTauriStorageCleanup to tauri/utils
