+++ application/ui/src/features/dataset/import-export/import-jobs-list/import-jobs-list.component.tsx
@@ -36,13 +36,6 @@ export const ImportJobsList = () => {
                 { params: { path: { project_id: projectId } } },
             ]),
         });
-        queryClient.invalidateQueries({
-            queryKey: getQueryKey([
-                'get',
-                '/api/projects/{project_id}/dataset/items',
-                { params: { path: { project_id: projectId } } },
-            ]),
-        });
     };
 
     const handleOpen = (stagedDatasetId: string) => {
