+++ application/ui/src/tauri/utils/storage-cleanup-registry/storage-cleanup-registry.test.ts
