+++ application/ui/src/tauri/utils/storage-cleanup-registry/storage-cleanup-registry.ts
@@ -1,9 +1,9 @@
 // Copyright (C) 2026 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { EXPORT_DATASET_PREFIX } from './use-export-dataset.hook';
-import { IMPORT_DATASET_AS_NEW_PROJECT_KEY } from './use-import-dataset-as-new-project.hook';
-import { IMPORT_DATASET_TO_PROJECT_PREFIX } from './use-import-dataset-to-project.hook';
+import { EXPORT_DATASET_PREFIX } from '../../../hooks/storage/use-export-dataset.hook';
+import { IMPORT_DATASET_AS_NEW_PROJECT_KEY } from '../../../hooks/storage/use-import-dataset-as-new-project.hook';
+import { IMPORT_DATASET_TO_PROJECT_PREFIX } from '../../../hooks/storage/use-import-dataset-to-project.hook';
 
 const DATASET_STORAGE_PREFIXES = [EXPORT_DATASET_PREFIX, IMPORT_DATASET_TO_PROJECT_PREFIX];
 const DATASET_STORAGE_KEYS = [IMPORT_DATASET_AS_NEW_PROJECT_KEY];
